+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/block/processor/model/GetWalletSummaryResult.kt
@@ -11,6 +11,8 @@ internal sealed class GetWalletSummaryResult {
         val walletSummary: WalletSummary
     ) : GetWalletSummaryResult() {
         fun scanProgressPercentDecimal() = PercentDecimal(walletSummary.scanProgress.getSafeRatio())
+        fun recoveryProgressPercentDecimal() =
+            walletSummary.recoveryProgress?.getSafeRatio()?.let { ratio -> PercentDecimal(ratio) }
     }
 
     data object None : GetWalletSummaryResult()