+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/model/WalletSummary.kt
@@ -9,6 +9,7 @@ internal data class WalletSummary(
     val chainTipHeight: BlockHeight,
     val fullyScannedHeight: BlockHeight,
     val scanProgress: ScanProgress,
+    val recoveryProgress: RecoveryProgress?,
     val nextSaplingSubtreeIndex: UInt,
     val nextOrchardSubtreeIndex: UInt
 ) {
@@ -23,6 +24,7 @@ internal data class WalletSummary(
                 chainTipHeight = BlockHeight(jni.chainTipHeight),
                 fullyScannedHeight = BlockHeight(jni.fullyScannedHeight),
                 scanProgress = ScanProgress.new(jni),
+                recoveryProgress = RecoveryProgress.new(jni),
                 nextSaplingSubtreeIndex = jni.nextSaplingSubtreeIndex.toUInt(),
                 nextOrchardSubtreeIndex = jni.nextOrchardSubtreeIndex.toUInt()
             )